Move ParticleType to its own file
